+++ packages/hightide-web/stories/User Interaction/Input/FlexibleDateTimeInput.stories.tsx
@@ -56,6 +56,10 @@ export const flexibleDateTimeInput: Story = {
         setIs24HourFormat: () => {},
         timeZone: timeZone || undefined,
         setTimeZone: () => {},
+        startingWeekday: 'monday',
+        setStartingWeekday: () => {},
+        calendarType: 'Gregorian',
+        setCalendarType: () => {},
         isInitialized: true,
       }}>
         <div className="flex-col-2 w-full max-w-md">