+++ packages/hightide-native/CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.10.2] - 2026-09-21
+
+### Added
+
+- Localization re-exports `Weekday` and `CalendarType`
+
+### Changed
+
+- Default `de-DE` and `en-US` locale info includes `defaultIs24HourFormat`, `defaultStartingWeekday`, and `defaultCalendarType`
+
 ## [0.10.1] - 2026-09-18
 
 ### Changed